be/requirements.txt
starlette
python-multipart
onnxruntime
fastapi
gunicorn
pydantic
pillow
imagehash
uvicorn[standard]
